🎨 use accentColor instaed of @color/oablue
This change will replace references to oablue with the accentColor
attribute. The main adavntage of using a style attribute here is that it
allows theming. e.g. If we choose so we could use this for slightly
different accentColors in day and night theme. IMHO it's also easier
collaborating when using standard nomenclature form Androids/Material
Designs color system.