📏 transparent status bar
Reworked the activity layouts for transparent status-bar support. (It's
a little quirky to get support for that on Android. Also the
navigation-bar will unconditionally become transparent when setting the
status-bar to transparent. So I tried to take care of that as well.)

Also removed `override dispatchTouchEvent` code duplication, by moving
it into `BaseActivity` and making sure all Activites in the app inherit
from that class.