⏱️  limit database progreess writes to 7/sec (#522)
I couldn't reproduce the crash described in #522. But I noticed some
severe performance hits during uploads to excessive write and read
operations on the database. (up to hundrets per second.) This occurs
because the upload progress in persisted in the database with very high
frequency. I've added a simple safeguard to reduce the frequency here to
~7 times per second.

This is a stop-gap measure and it might make sense to reconsider the
architecture of continously writing download progress to the database.
Most other apps display progress as bundle-value attached to a
broadcast.