🧋 request notification permission
JobService needs to create a notification for starting a user initiates
job. So I think we probably want to request notification permission for
displaying it.

On a side note: I think it would be technically possible to start a
notification without requesting permission first. This would effectively
allow us to run uploads in a user intiated background job, without a
notification.