🪄 add feedback to gdrive auth
With this change Save will give some visual feedback to users when
Google Drive authentication starts.  In the simple from of disabling the
Buttons while authentication is running.  Also running 2 authentication
attempts at the same time is not supported by our code, so this prevents
users to get into that error state.