feat(flavor): move play integrity into flavor build
some other bug fixes with exception handling

ignoring apps with no integrity for now, useful for proof or logging

separating the build source int 'app/src/play'.  'app/src/fdroid' is
  currently unused, but will not include the integrity feature.

added structure to dynamically load features which can be reused by
other features.  Should be a different branch, maybe with more time.

currently features are specified in the manifest with a 'meta-data' property with the
key as the class, and a value of 'feature'.

the buildsrc manifest takes advantage of manifest merging.

when the dynamic features are found, their koin modules are loaded, and
they run 'onLoad' (or 'onUnload') with the main activity.